doc-hidden submacros, strict feature, various fixes/improvements
