+++ src/lib.rs
@@ -141,13 +141,12 @@
 //! impl Data {
 //!     /// Adds a file, updates the clients concerned.
 //!     pub fn add_file(&mut self, file: FileInfo) -> File {
-//!         let idx = self.files.next_index();
+//!         let idx = self.files.push(file);
+//!         let file = &self.files[idx];
 //!         for client in &file.clients {
 //!             let is_new = self.clients[*client].files.insert(idx);
 //!             debug_assert! { is_new }
 //!         }
-//!         let nu_idx = self.files.push(file);
-//!         debug_assert_eq! { idx, nu_idx }
 //!         idx
 //!     }
 //!
@@ -166,35 +165,80 @@
 //! [`examples::clients`]: examples/clients/index.html (clients example)
 //! [clients src]: examples/clients.rs.html (Code of the clients example)
 
-/// Wraps a `usize` into a struct (zero-cost). Also generates the relevant collections indexed by
-/// the wrapper.
-///
-/// See the [module-level documentation](index.html) for more.
+mod map;
+
+/// Discards its input if the `strict` feature is active.
 #[macro_export]
-macro_rules! new {
-    // Btree set (internal).
-    ( @internal $t:ident, $(#[$cmt:meta])* btree set: $set:ident $($tail:tt)* ) => (
-        $(#[$cmt])*
+#[doc(hidden)]
+#[cfg(feature = "strict")]
+macro_rules! non_strict {
+    ( $($stuff:tt)* ) => {};
+}
+/// Discards its input if the `strict` feature is active.
+#[macro_export]
+#[doc(hidden)]
+#[cfg(not(feature = "strict"))]
+macro_rules! non_strict {
+    (
+        #[doc = $doc:literal]
+        $($tail:tt)*
+    ) => {
+        #[doc = concat!("[`non_strict`] ", $doc)]
+        $($tail)*
+    };
+    (
+        $($item:item)*
+    ) => {
+        $(
+            #[doc = "[`non_strict`] "]
+            $item
+        )*
+    };
+}
+
+/// Generates an alias type for [`std::collections::BTreeSet`] of indices.
+#[macro_export]
+#[doc(hidden)]
+macro_rules! btree_set_codegen {
+    { $t:ident,
+        $(#[$meta:meta])*
+        $set:ident $($tail:tt)*
+    } => {
+        $(#[$meta])*
         pub type $set = std::collections::BTreeSet<$t> ;
-        $crate::new!{ @internal $t $($tail)* }
-    ) ;
+        $crate::handle!{ $t $($tail)* }
+    };
+}
 
-    // Btree map (internal).
-    ( @internal $t:ident, $(#[$cmt:meta])* btree map: $map:ident $($tail:tt)* ) => (
-        $(#[$cmt])*
+/// Generates an alias type for [`std::collections::BTreeMap`] of indices.
+#[macro_export]
+#[doc(hidden)]
+macro_rules! btree_map_codegen {
+    { $t:ident,
+        $(#[$meta:meta])*
+        $map:ident $($tail:tt)*
+    } => {
+        $(#[$meta])*
         pub type $map<T> = std::collections::BTreeMap<$t, T> ;
-        $crate::new!{ @internal $t $($tail)* }
-    ) ;
+        $crate::handle!{ $t $($tail)* }
+    };
+}
 
-    // Range (internal).
-    ( @internal $t:ident, $(#[$cmt:meta])* range: $range:ident $($tail:tt)* ) => (
+/// Generates an alias type for [`std::collections::BTreeMap`] of indices.
+#[macro_export]
+#[doc(hidden)]
+macro_rules! range_codegen {
+    { $t:ident,
+        $(#[$meta:meta])*
+        $range:ident $($tail:tt)*
+    } => {
         impl $t {
-            /// Creates a range between two indexes (upper bound is exclusive).
+            /// Creates a range between two indices (upper bound is exclusive).
             pub fn up_to(self, end : $t) -> $range {
                 $range { start: self, end }
             }
         }
-        $(#[$cmt])*
+        $(#[$meta])*
         #[derive(Debug)]
         pub struct $range {
             start: $t,
@@ -210,7 +254,7 @@ macro_rules! new {
             }
             /// Creates a range from `0` to something.
             pub fn zero_to<T: std::convert::Into<$t>>(end: T) -> Self {
-                $range { start: 0.into(), end: end.into() }
+                $range { start: $t { val: 0 }, end: end.into() }
             }
         }
         impl std::iter::Iterator for $range {
@@ -223,455 +267,86 @@ macro_rules! new {
                 }
             }
         }
-        $crate::new!{ @internal $t $($tail)* }
-    ) ;
+        $crate::handle!{ $t $($tail)* }
+    };
+}
 
-    // Map: vector indexed by `$t` (internal).
+/// Handles some user input and decides what to do.
+#[macro_export]
+#[doc(hidden)]
+macro_rules! handle {
+    { $t:ident, $(#[$meta:meta])* btree set: $($tail:tt)* } => {
+        $crate::btree_set_codegen! { $t, $(#[$meta])* $($tail)* }
+    };
+    { $t:ident, $(#[$meta:meta])* btree map: $($tail:tt)* } => {
+        $crate::btree_map_codegen! { $t, $(#[$meta])* $($tail)* }
+    };
+    { $t:ident, $(#[$meta:meta])* range: $($tail:tt)* } => {
+        $crate::range_codegen! { $t, $(#[$meta])* $($tail)* }
+    };
+    { $t:ident, $(#[$meta:meta])* map: $($tail:tt)* } => {
+        $crate::map_codegen! { $t, $(#[$meta])* $($tail)* }
+    };
+    { $t:ident $(,)? } => {};
+}
+
+/// Wraps a `usize` into a struct (zero-cost). Also generates the relevant collections indexed by
+/// the wrapper.
+///
+/// See the [module-level documentation](index.html) for more.
+#[macro_export]
+macro_rules! new {
     (
-        @internal $t:ident, $(#[$cmt:meta])*
-        map: $map:ident with iter: $iter:ident
+        $(#[$meta:meta])*
+        $t:ident
         $($tail:tt)*
     ) => (
-        $(#[$cmt])*
-        #[derive(Debug, PartialOrd, Ord)]
-        pub struct $map<T> {
-            vec: Vec<T>
-        }
-        impl<T> Default for $map<T> {
-            fn default() -> Self { Self::new() }
-        }
-        impl<T: Clone> Clone for $map<T> {
-            fn clone(& self) -> Self {
-                $map { vec: self.vec.clone() }
-            }
-        }
-        impl<T> $map<T> {
-            /// Creates an empty map from an existing one.
-            #[inline]
-            pub const fn of(vec: Vec<T>) -> Self {
-                $map { vec: vec }
-            }
-            /// Creates an empty map.
-            #[inline]
-            pub const fn new() -> Self {
-                $map { vec: Vec::new() }
-            }
-            /// Creates an empty map with some capacity.
-            #[inline]
-            pub fn with_capacity(capacity: usize) -> Self {
-                $map { vec: Vec::with_capacity(capacity) }
-            }
-
-            /// Reserves some space for the map.
-            #[inline]
-            pub fn reserve(&mut self, capa: usize) {
-                self.vec.reserve(capa)
-            }
-
-            /// Retrieves an entry in the map.
-            #[inline]
-            pub fn get(&self, uid: $t) -> Option<&T> {
-                self.vec.get(uid.get())
-            }
-            /// Retrieves an entry in the map.
-            #[inline]
-            pub fn get_mut(&mut self, uid: $t) -> Option<&mut T> {
-                self.vec.get_mut(uid.get())
-            }
-
-            /// Retrieves the last entry in the map.
-            #[inline]
-            pub fn last(&self) -> Option<&T> {
-                self.vec.last()
-            }
-            /// Retrieves the last entry in the map.
-            #[inline]
-            pub fn last_mut(&self) -> Option<&T> {
-                self.vec.last()
-            }
-
-            /// Number of elements in the map.
-            #[inline]
-            pub fn len(& self) -> usize {
-                self.vec.len()
-            }
-            /// Capacity of the map.
-            #[inline]
-            pub fn capacity(& self) -> usize {
-                self.vec.capacity()
-            }
-
-            /// The next free index (wrapped `self.len()`).
-            #[inline]
-            pub fn next_index(& self) -> $t {
-                self.len().into()
-            }
-            /// The last index in the map.
-            #[inline]
-            pub fn last_index(& self) -> Option<$t> {
-                let len = self.len();
-                if len > 0 { Some((len - 1).into()) } else { None }
-            }
-
-            /// Pushes an element.
-            #[inline]
-            pub fn push(&mut self, elem: T) -> $t {
-                let idx = self.next_index();
-                self.vec.push(elem);
-                idx
-            }
-            /// Pops an element.
-            ///
-            /// This function is unsafe for the logics of safe indices. This function voids indices
-            /// previously created (indices for the last element on entry) and should be used with
-            /// great care.
-            #[inline]
-            pub fn pop(&mut self) -> Option<T> {
-                self.vec.pop()
-            }
-
-            /// Clears a map.
-            #[inline]
-            pub fn clear(&mut self) {
-                self.vec.clear()
-            }
-
-            /// Iterates over the elements.
-            #[inline]
-            pub fn iter(& self) -> std::slice::Iter<T> {
-                self.vec.iter()
-            }
-            /// Iterates over the elements with the index.
-            #[inline]
-            pub fn index_iter<'a>(&'a self) -> $iter<&'a $map<T>>
-            where T: 'a {
-                $iter::mk_ref(self)
-            }
-            /// Iterates over the elements with the index, mutable version.
-            #[inline]
-            pub fn index_iter_mut<'a>(&'a mut self) -> $iter<
-                std::slice::IterMut<'a, T>
-            >
-            where T: 'a {
-                $iter::mk_ref_mut(self)
-            }
-            /// Iterates over the elements with the index.
-            #[inline]
-            pub fn into_index_iter(self) -> $iter<$map<T>> {
-                $iter::new(self)
-            }
-            /// Iterates over the elements (mutable version).
-            #[inline]
-            pub fn iter_mut(&mut self) -> std::slice::IterMut<T> {
-                self.vec.iter_mut()
-            }
-
-            /// Shrinks the capacity as much as possible.
-            #[inline]
-            pub fn shrink_to_fit(&mut self) {
-                self.vec.shrink_to_fit()
-            }
-            /// Swap from `Vec`.
-            #[inline]
-            pub fn swap(&mut self, a: $t, b: $t) {
-                self.vec.swap(* a, *b)
-            }
-            /// Swap remove from `Vec`.
-            ///
-            /// This function is unsafe for the logics of safe indices. This function voids indices
-            /// previously created (indices for the last element on entry) and should be used with
-            /// great care.
-            #[inline]
-            pub fn swap_remove(&mut self, idx: $t) -> T {
-                self.vec.swap_remove(* idx)
-            }
-
-            /// Splits the map into the elements before and after some index.
-            ///
-            /// More precisely, returns a tuple of
-            ///
-            /// - an iterator over the elements *before* `idx`,
-            /// - the element at position `idx`, and
-            /// - an iterator over the elements *after* `idx`.
-            #[inline]
-            pub fn split(&self, idx: $t) -> (impl Iterator<Item = &T>, &T, impl Iterator<Item = &T>) {
-                let before = self.vec[0..idx.val].iter();
-                let after = if idx.val < self.vec.len() {
-                    self.vec[idx.val + 1 ..].iter()
-                } else {
-                    self.vec[0..0].iter()
-                };
-                (before, &self.vec[idx.val], after)
-            }
-        }
-
-        impl<T: Clone> $map<T> {
-            /// Creates an empty vector with some capacity.
-            #[inline]
-            pub fn of_elems(elem: T, size: usize) -> Self {
-                $map { vec: vec![ elem ; size ] }
-            }
-        }
-        impl<T: PartialEq> PartialEq for $map<T> {
-            fn eq(& self, other: & Self) -> bool {
-                self.vec.eq( & other.vec )
-            }
-        }
-
-        impl<T: Eq> Eq for $map<T> {}
-
-        impl<T> std::convert::From< Vec<T> > for $map<T> {
-            fn from(vec: Vec<T>) -> Self {
-                $map { vec }
-            }
-        }
-        impl<T> std::iter::IntoIterator for $map<T> {
-            type Item = T ;
-            type IntoIter = std::vec::IntoIter<T> ;
-            fn into_iter(self) -> std::vec::IntoIter<T> {
-                self.vec.into_iter()
-            }
-        }
-        impl<'a, T> std::iter::IntoIterator for &'a $map<T> {
-            type Item = &'a T ;
-            type IntoIter = std::slice::Iter<'a, T> ;
-            fn into_iter(self) -> std::slice::Iter<'a, T> {
-                self.iter()
-            }
-        }
-        impl<'a, T> std::iter::IntoIterator for &'a mut $map<T> {
-            type Item = &'a mut T ;
-            type IntoIter = std::slice::IterMut<'a, T> ;
-            fn into_iter(self) -> std::slice::IterMut<'a, T> {
-                self.iter_mut()
-            }
-        }
-        impl<T> std::iter::FromIterator<T> for $map<T> {
-            fn from_iter<
-                I: std::iter::IntoIterator<Item = T>
-            >(iter: I) -> Self {
-                $map { vec: iter.into_iter().collect() }
-            }
-        }
-        impl<T> std::ops::Index<$t> for $map<T> {
-            type Output = T ;
-            fn index(& self, index: $t) -> & T {
-                & self.vec[ index.get() ]
-            }
-        }
-        impl<T> std::ops::Index<std::ops::RangeFrom<$t>> for $map<T> {
-            type Output = [T];
-            fn index(& self, std::ops::RangeFrom { start }: std::ops::RangeFrom<$t>) -> &[T] {
-                & self.vec[ start.get() .. ]
-            }
-        }
-        impl<T> std::ops::Index<std::ops::Range<$t>> for $map<T> {
-            type Output = [T];
-            fn index(& self, std::ops::Range { start, end }: std::ops::Range<$t>) -> &[T] {
-                & self.vec[ start.get() .. end.get() ]
-            }
-        }
-        impl<T> std::ops::Index<std::ops::RangeInclusive<$t>> for $map<T> {
-            type Output = [T];
-            fn index(& self, range: std::ops::RangeInclusive<$t>) -> &[T] {
-                & self.vec[ range.start().get() ..= range.end().get() ]
-            }
-        }
-        impl<T> std::ops::Index<std::ops::RangeFull> for $map<T> {
-            type Output = [T];
-            fn index(& self, _: std::ops::RangeFull) -> &[T] {
-                & self.vec[..]
-            }
-        }
-        impl<T> std::ops::Index<std::ops::RangeTo<$t>> for $map<T> {
-            type Output = [T];
-            fn index(& self, std::ops::RangeTo { end }: std::ops::RangeTo<$t>) -> &[T] {
-                & self.vec[..end.get()]
-            }
-        }
-        impl<T> std::ops::Index<std::ops::RangeToInclusive<$t>> for $map<T> {
-            type Output = [T];
-            fn index(
-                & self,
-                std::ops::RangeToInclusive { end }: std::ops::RangeToInclusive<$t>
-            ) -> &[T] {
-                & self.vec[..=end.get()]
-            }
-        }
-        impl<T> std::ops::IndexMut<$t> for $map<T> {
-            fn index_mut(&mut self, index: $t) -> &mut T {
-                &mut self.vec[ index.get() ]
-            }
-        }
-        impl<T> std::ops::Index<
-            std::ops::Range<usize>
-        > for $map<T> {
-            type Output = [T] ;
-            fn index(& self, index: std::ops::Range<usize>) -> & [T] {
-                self.vec.index(index)
-            }
-        }
-        impl<T> std::ops::Index<
-          std::ops::RangeInclusive<usize>
-        > for $map<T> {
-          type Output = [T] ;
-          fn index(& self, index: std::ops::RangeInclusive<usize>) -> & [T] {
-            self.vec.index(index)
-          }
-        }
-        impl<T> std::ops::Index<
-            std::ops::RangeFrom<usize>
-        > for $map<T> {
-            type Output = [T] ;
-            fn index(& self, index: std::ops::RangeFrom<usize>) -> & [T] {
-                self.vec.index(index)
-            }
-        }
-        impl<T> std::ops::Index<
-            std::ops::RangeTo<usize>
-        > for $map<T> {
-            type Output = [T] ;
-            fn index(& self, index: std::ops::RangeTo<usize>) -> & [T] {
-                self.vec.index(index)
-            }
-        }
-        // impl<T> std::ops::Index<
-        //   std::ops::RangeToInclusive<usize>
-        // > for $map<T> {
-        //   type Output = [T] ;
-        //   fn index(& self, index: std::ops::RangeToInclusive<usize>) -> & [T] {
-        //     self.vec.index(index)
-        //   }
-        // }
-        impl<T> std::ops::Deref for $map<T> {
-            type Target = Vec<T> ;
-            fn deref(& self) -> & Vec<T> {
-                & self.vec
-            }
-        }
-        /// Structure allowing to iterate over the elements of a map and their
-        /// index.
-        #[derive(Clone)]
-        pub struct $iter<T> {
-            cursor: $t,
-            map: T,
-        }
-        impl<'a, T> $iter<&'a $map<T>> {
-            /// Creates an iterator starting at 0.
-            const fn mk_ref(map: &'a $map<T>) -> Self {
-                $iter { cursor: $t::zero(), map: map }
-            }
-        }
-        impl<'a, T: 'a> std::iter::Iterator for $iter<&'a $map<T>> {
-            type Item = ($t, &'a T) ;
-            fn next(&mut self) -> Option< ($t, &'a T) > {
-                if self.cursor >= self.map.len() {
-                    None
-                } else {
-                    let res = (self.cursor, & self.map[self.cursor]) ;
-                    self.cursor.inc() ;
-                    Some(res)
-                }
-            }
-        }
-        impl<'a, T: 'a> $iter<std::slice::IterMut<'a, T>> {
-            /// Creates an iterator starting at 0, mutable version.
-            fn mk_ref_mut(map: &'a mut $map<T>) -> Self {
-                $iter { cursor: $t::zero(), map: map.vec.iter_mut() }
-            }
-        }
-        impl<'a, T: 'a> std::iter::Iterator for $iter<
-            std::slice::IterMut<'a, T>
-        > {
-            type Item = ($t, &'a mut T) ;
-            fn next(&mut self) -> Option< ($t, &'a mut T) > {
-                self.map.next().map(
-                    |res| {
-                        let index = self.cursor ;
-                        self.cursor.inc() ;
-                        (index, res)
-                    }
-                )
-            }
-        }
-        impl<T> $iter<$map<T>> {
-            /// Creates an iterator starting at 0.
-            fn new(mut map: $map<T>) -> Self {
-                map.vec.reverse() ;
-                $iter { cursor: $t::zero(), map: map }
-            }
-        }
-        impl<T> std::iter::Iterator for $iter<$map<T>> {
-            type Item = ($t, T) ;
-            fn next(&mut self) -> Option< ($t, T) > {
-                if let Some(elem) = self.map.pop() {
-                    let res = (self.cursor, elem) ;
-                    self.cursor.inc() ;
-                    Some(res)
-                } else {
-                    None
-                }
-            }
-        }
-        $crate::new!{ @internal $t $($tail)* }
-    ) ;
-
-    // Terminal case (internal).
-    ( @internal $t:ident $(,)? ) => () ;
-
-    // Entry point.
-    (
-        $(#[$cmt:meta])* $t:ident
-        $($tail:tt)*
-    ) => (
-        $(#[$cmt])*
+        $(#[$meta])*
         #[derive(Debug, Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Hash)]
         pub struct $t {
             val: usize
         }
 
         impl $t {
-            /// Wraps an int.
-            #[inline]
-            pub const fn new(val: usize) -> Self {
-                $t { val: val }
+            $crate::non_strict! {
+                /// Wraps a [`usize`].
+                #[inline]
+                pub const fn new(val: usize) -> Self {
+                    $t { val: val }
+                }
             }
-            /// Zero.
-            #[inline]
-            pub const fn zero() -> Self {
-                $t { val: 0 }
+            $crate::non_strict! {
+                /// Zero.
+                #[inline]
+                pub const fn zero() -> Self {
+                    $t { val: 0 }
+                }
             }
-            /// One.
-            #[inline]
-            pub const fn one() -> Self {
-                $t { val: 1 }
+            $crate::non_strict! {
+                /// One.
+                #[inline]
+                pub const fn one() -> Self {
+                    $t { val: 1 }
+                }
             }
-            /// Accessor.
+            $crate::non_strict! {
+                /// Increments the int.
+                #[inline]
+                pub fn inc(&mut self) {
+                    self.val += 1
+                }
+            }
+            $crate::non_strict! {
+                /// Decrements the int.
+                #[inline]
+                pub fn dec(&mut self) {
+                    self.val -= 1
+                }
+            }
+            /// Underlying index accessor.
             #[inline]
             pub const fn get(& self) -> usize {
                 self.val
-            }
-            /// Increments the int.
-            #[inline]
-            pub fn inc(&mut self) {
-                self.val += 1
-            }
-            /// Decrements the int.
-            #[inline]
-            pub fn dec(&mut self) {
-                self.val -= 1
-            }
-        }
-        impl std::convert::From<usize> for $t {
-            #[inline]
-            fn from(val: usize) -> Self {
-                $t::new(val)
-            }
-        }
-        impl<'a> std::convert::From<&'a usize> for $t {
-            #[inline]
-            fn from(val: &'a usize) -> Self {
-                $t::new(* val)
             }
         }
         impl std::convert::Into<usize> for $t {
@@ -686,24 +361,12 @@ macro_rules! new {
                 self.val
             }
         }
-        impl<T: std::convert::Into<usize>> std::ops::AddAssign<T> for $t {
-            #[inline]
-            fn add_assign(&mut self, rhs: T) {
-                self.val += rhs.into()
-            }
-        }
         impl<T: std::convert::Into<usize>> std::ops::Add<T> for $t {
             type Output = $t ;
             #[inline]
             fn add(mut self, rhs: T) -> $t {
                 self.val += rhs.into() ;
                 self
-            }
-        }
-        impl Default for $t {
-            #[inline]
-            fn default() -> Self {
-                Self::zero()
             }
         }
         impl std::ops::Deref for $t {
@@ -733,7 +396,33 @@ macro_rules! new {
                 self.val.partial_cmp(int)
             }
         }
-        $crate::new!{ @internal $t $($tail)* }
+        $crate::non_strict! {
+            impl std::convert::From<usize> for $t {
+                #[inline]
+                fn from(val: usize) -> Self {
+                    $t::new(val)
+                }
+            }
+            impl<'a> std::convert::From<&'a usize> for $t {
+                #[inline]
+                fn from(val: &'a usize) -> Self {
+                    $t::new(* val)
+                }
+            }
+            impl<T: std::convert::Into<usize>> std::ops::AddAssign<T> for $t {
+                #[inline]
+                fn add_assign(&mut self, rhs: T) {
+                    self.val += rhs.into()
+                }
+            }
+            impl Default for $t {
+                #[inline]
+                fn default() -> Self {
+                    Self::zero()
+                }
+            }
+        }
+        $crate::handle!{ $t $($tail)* }
     ) ;
 }
